requirements.txt
# Устанавливаем локальный crewai из скачанного репозитория:
# pip install -e ../lib/crewai
# pip install -e ../lib/crewai-tools

# ── Core ──────────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# ── LLM ───────────────────────────────────────────────────────────────────────
openai>=1.50.0

# ── PostgreSQL (Supabase или self-hosted) ─────────────────────────────────────
asyncpg>=0.29.0          # async PostgreSQL для FastAPI (память диалогов)
psycopg2-binary>=2.9.0   # sync PostgreSQL для CrewAI tools (vector store)

# ── Observability ─────────────────────────────────────────────────────────────
langfuse>=2.0.0          # трассировка агентов (опционально, можно удалить)

# ── HTTP ─────────────────────────────────────────────────────────────────────
httpx>=0.27.0
